Remove PoolableString again
