+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/DebugComponent.kt
@@ -2,7 +2,7 @@ package korlibs.korge.fleks.components
 
 
 import com.github.quillraven.fleks.*
-import korlibs.korge.fleks.utils.componentPool.*
+import korlibs.korge.fleks.utils.CloneableComponent
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
