+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/EntityLinkComponent.kt
@@ -2,7 +2,6 @@ package korlibs.korge.fleks.components
 
 import com.github.quillraven.fleks.*
 import korlibs.korge.fleks.utils.*
-import korlibs.korge.fleks.utils.componentPool.*
 import kotlinx.serialization.*
