+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/NoisyMoveComponent.kt
@@ -3,7 +3,6 @@ package korlibs.korge.fleks.components
 import com.github.quillraven.fleks.*
 import korlibs.korge.fleks.utils.*
 import korlibs.korge.fleks.components.TweenPropertyComponent.*
-import korlibs.korge.fleks.utils.componentPool.*
 import korlibs.math.interpolation.*
 import kotlinx.serialization.*
