+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/SpriteComponent.kt
@@ -5,7 +5,6 @@ import korlibs.image.format.*
 import korlibs.image.format.ImageAnimation.Direction.*
 import korlibs.korge.fleks.assets.AssetStore
 import korlibs.korge.fleks.utils.*
-import korlibs.korge.fleks.utils.componentPool.*
 import korlibs.time.*
 import kotlinx.serialization.*
