+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/components/TiledLevelMapComponent.kt
@@ -1,7 +1,7 @@
 package korlibs.korge.fleks.components
 
 import com.github.quillraven.fleks.*
-import korlibs.korge.fleks.utils.componentPool.*
+import korlibs.korge.fleks.utils.CloneableComponent
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
