+++ korge-fleks/src/commonMain/kotlin/korlibs/korge/fleks/utils/InjectableConfigurationExt.kt
@@ -3,13 +3,19 @@ package korlibs.korge.fleks.utils
 import com.github.quillraven.fleks.*
 import korlibs.korge.fleks.components.Collision.Companion.addCollisionComponentPool
 import korlibs.korge.fleks.components.LevelMap.Companion.addLevelMapComponentPool
-import korlibs.korge.fleks.utils.poolableData.Point.Companion.addPointDataPool
+import korlibs.korge.fleks.components.data.Point.Companion.addPointDataPool
 
+/**
+ * Add here all component pools which are used in the game.
+ */
 fun InjectableConfiguration.addKorgeFleksComponentPools(preAllocate: Int = 0) {
     addCollisionComponentPool(preAllocate)
     addLevelMapComponentPool(preAllocate)
 }
 
+/**
+ * Add here all data pools which are used in the game.
+ */
 fun InjectableConfiguration.addKorgeFleksDataPools(preAllocate: Int = 0) {
     addPointDataPool(preAllocate)
 }
